chore: extract sql commenting logic to common package
