+++ packages/opentelemetry-sql-common/.eslintignore
@@ -1,0 +1,1 @@
+build
