+++ packages/opentelemetry-sql-common/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  env: {
+    mocha: true,
+    node: true,
+  },
+  ...require('../../eslint.config.js'),
+};
