+++ plugins/node/opentelemetry-instrumentation-pg/package.json
@@ -77,6 +77,7 @@
     "@opentelemetry/core": "^1.8.0",
     "@opentelemetry/instrumentation": "^0.39.1",
     "@opentelemetry/semantic-conventions": "^1.0.0",
+    "@opentelemetry/sql-common": "^0.39.0",
     "@types/pg": "8.6.1",
     "@types/pg-pool": "2.0.3"
   },
